changed more simulation-specific types (life, energy, max-life, etc.) from int to float
git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@1657 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Simulation/Properties/Environment/ExplosionControllerProperty.cs
@@ -51,7 +51,7 @@ namespace ProjectMagma.Simulation
                 // apply damage to player
                 if (explosion.HasAttribute("damage"))
                 {
-                    other.SetInt("health", other.GetInt("health") - explosion.GetInt("damage"));
+                    other.SetFloat("health", other.GetFloat("health") - explosion.GetFloat("damage"));
                 }
                 if(explosion.HasAttribute("freeze_time"))
                 {
